Added UI for status, fixed a bug in damage calculation
Damage calculation was erroneously using an Integer for the stat modifier multiplication which resulted in the stat being set to 0 when using debuffed values.

Minor error, but pretty significant effect. Good thing I caught it relatively soon.

Also added a 100% Paralysis chance move, for testing.

Status effects are now applied by moves and display properly on the HUD, but currently have no actual effects.
The exception for this is Burn, which does not apply end of turn damage, but DOES affect the damage dealt by physical moves.