Added a preload scene and audio manager, fixed tooltip bug
The preload scene initializes the manager objects required for the game, and should always be run first.

The tooltip bug in question was hard to reproduce but easily fixed. Sometimes the tooltips for mousing over moves in battle were not properly removed, causing them to stack up.
Now before a tooltip is created, a check is performed to see if one already exists, and deletes it if it does.