Fixed a bug involving end of turn status effects
End of turn status effects that dealt damage (poison and burn) were dealing damage correctly, but not updating the HP bars to match.
Also, a Faint Check was not properly performed afterwards, so a monster that was brought to 0 hit points by poison or burn would not properly faint until after the next battle action.